requirements.txt
urllib3==1.26.6
requests==2.26.0
jsonpath==0.82
music_tag==0.4.3
PyYAML==5.4.1
tqdm==4.61.2
